# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Beckbest Bridal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
  /bb-portal/products:
    get:
      operationId: adminListProducts
      tags: [portal]
      summary: Admin — list all products
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: adminCreateProduct
      tags: [portal]
      summary: Admin — create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /bb-portal/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [portal]
      summary: Admin — update product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: adminDeleteProduct
      tags: [portal]
      summary: Admin — delete product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        images:
          type: array
          items: { type: string }
        inStock: { type: boolean }
        featured: { type: boolean }
        sku: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, name, price, images, inStock, featured, createdAt]
    ProductList:
      type: object
      properties:
        products:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        total: { type: integer }
      required: [products, total]
    ProductInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        compareAtPrice: { type: number }
        categoryId: { type: integer }
        images:
          type: array
          items: { type: string }
        inStock: { type: boolean }
        featured: { type: boolean }
        sku: { type: string }
      required: [name, price]
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        compareAtPrice: { type: number }
        categoryId: { type: integer }
        images:
          type: array
          items: { type: string }
        inStock: { type: boolean }
        featured: { type: boolean }
        sku: { type: string }
...